[gnocchi-14] Adding pairwise sample similarity comparison.
This uses the DIMSUM algorithm to compute the cosine similarity between two
samples in a variant call dataset. To do this, we discard all non-alt calls,
and count the number of alt calls per sample per alt allele, and then we call
the DIMSUM algorithm, using the alt call count per site as a sparse matrix row.