[gnocchi-18] Port Gnocchi over to the Spark SQL DataFrame/Dataset API
Because Spark SQL presents better opportunities for optimization in the future,
I have moved Gnocchi over to Spark SQL. Several functions are not working
correctly right now; there seems to be an issue with Broadcast JOINs that
crops up when running regressPhenotypes. I plan to resolve that before merging
this work. Resolves #18.