[gnocchi-19] Add PCA.
This commit adds code to run PCA on genotypes. This code transposes the genotype
matrix and then uses SVD to compute the PCA rotations. This allows us to process
all of the data, instead of filtering down to a genomic range and then running
PCA. Additionally, I have factored the matrix creation code out of the
SampleSimilarity code. Resolves #19.